Move DailySurveyRetriever into its own file. Remove logger.
